Add files via upload
This application was done as a school project in a group of two. The main function of this application is to give information of a desired bands timetable and the location of their upcoming concerts or performances. Additionally the user can get real time navigation to a destination via google maps that is linked to the actual application.